% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_mix_EM_15Jan2018_marg_lik.R
\name{BFmixture}
\alias{BFmixture}
\title{calculate BF with mixture prior.}
\usage{
BFmixture(obs.data, nvariants, low.bound = 10^(-2))
}
\arguments{
\item{obs.data}{N*2 data matrix, N is the sample size, 1st column is sum of rare variants for each individual, 2nd column is the disease status}

\item{nvariants}{number of sites of the region}

\item{low.bound}{lower bound}
}
\value{
Vector of 3 elements. 1st, probability of p=0; 2nd, precision parameter of beta distribution,3rd: BF
}
\description{
calculate BF with mixture prior.
}
\examples{

nsites = 500
ncase = 200
ncontrol = 200
maf = 0.005
rexp_param = 2

model1 = function(i) {
 
 p1 = rbinom( ncontrol  , nsites , maf   )
 p2 = rbinom( ncase     , nsites , maf )  + H1 * round(rexp(ncase,rexp_param))
 disease_status = c (  rep(0,ncontrol),rep(1,ncase)  )
 data = data.frame(c(p1,p2), disease_status)
 bf = BFmixture(data, nsites)[3]
 bf
}

nrep = 20

H1 = 0
bf1 = ( replicate(nrep,model1() ) )

H1 = 1
bf2 = ( replicate(nrep,model1() ) )

col = c( rep("blue",nrep) , rep("red", nrep))


plot(log(c(bf1,bf2)),col = col ,pch=20,cex=1.5)

}
