% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BF-package.r
\docType{package}
\name{rareBF-package}
\alias{rareBF}
\alias{rareBF-package}
\title{Methods for rare variant association.}
\description{
This packages implements methods for rare variant association. The following
two methods are provided:
BFmixture, BFbeta
}
