% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_mix_EM_15Jan2018_marg_lik.R
\name{BFmixture}
\alias{BFmixture}
\title{calculate BF with mixture prior.}
\usage{
BFmixture(obs.data, nvariants, low.bound = 10^(-2))
}
\arguments{
\item{obs.data}{N*2 data matrix, N is the sample size, 1st column is sum of rare variants for each individual, 2nd column is the disease status}

\item{nvariants}{number of sites of the region}

\item{low.bound}{lower bound}
}
\value{
Vector of 3 elements. 1st, probability of p=0; 2nd, precision parameter of beta distribution,3rd: BF
}
\description{
calculate BF with mixture prior.
}
